09122014
Completed combat, added labels for crits and gold reward. Still need to
add progression and gold spending.

+++ HeroTrainer_MythicPixel/HeroTrainer_MythicPixel/Warrior.cs
@@ -16,7 +16,7 @@ namespace HeroTrainer_MythicPixel
 		public float Luck{ get{return luck;} set{luck = value;} }
 		public float Haste{ get{return haste;} set{haste = value;} }
 		public float Opportunity{ get{return opportunity;} set{opportunity = value;} }
-		public float Gold{ get{return gold;} set{gold = value;} }
+		public int Gold{ get{return gold;} set{gold = value;} }
 		public string Name{ get{return name;} }
 		
 		private static string name;
@@ -25,7 +25,7 @@ namespace HeroTrainer_MythicPixel
 		private static float luck;
 		private static float haste;
 		private static float opportunity;
-		private static float gold;
+		private static int gold;
 		
 		private static SpriteUV		bob;
 		private static TextureInfo	bobInfo;
@@ -84,7 +84,7 @@ namespace HeroTrainer_MythicPixel
 					else if (line.IndexOf ('g') == 0)
 					{
 						line = line.TrimStart ('g');
-						gold = float.Parse (line);						
+						gold = int.Parse (line);						
 					}
 					else
 					{
